@@ -16,15 +16,16 @@
     "@nuxtjs/axios": "^5.6.0",
     "axios": "^0.19.0",
     "bootstrap-vue": "^2.0.4",
-    "nuxt": "^2.0.0"
+    "nuxt": "^2.10.1"
   },
   "devDependencies": {
     "@nuxtjs/style-resources": "^1.0.0",
-    "babel-eslint": "^10.0.1",
-    "eslint": "^4.19.1",
+    "@nuxtjs/vuetify": "^1.9.0",
+    "babel-eslint": "^10.0.3",
+    "eslint": "^6.5.1",
     "eslint-friendly-formatter": "^4.0.1",
-    "eslint-loader": "^2.1.1",
-    "eslint-plugin-vue": "^4.0.0",
+    "eslint-loader": "^3.0.2",
+    "eslint-plugin-vue": "^5.2.3",
     "node-sass": "^4.12.0",
     "sass-loader": "^8.0.0"
   }
